requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
idna==2.10
joblib==0.14.1
jsonpatch==1.28
jsonpointer==2.0
nltk==3.5
numpy==1.16.0
pandas==0.18.1
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.1
scipy==1.4.1
six==1.15.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.2.1
tornado==6.1
tqdm==4.57.0
urllib3==1.26.3
visdom==0.1.8.9
websocket-client==0.57.0
